dasbackend/pyproject.toml
[project]
name = "backend-task-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Ensar Hujdur",email = "ensar.hujdur196@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "logger (>=1.4,<2.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "databases (>=0.9.0,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
